requirements.txt
asgiref==3.2.3
boto3==1.12.0
botocore==1.15.0
Django==3.0.3
django-countries==5.5
django-localflavor==2.2
django-phone-field==1.8.0
django-storages==1.9.1
docutils==0.15.2
jmespath==0.9.4
Pillow==6.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-stdnum==1.12
pytz==2019.3
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
